// src/content/app/genome-browser/services/genome-browser-service/types/genomeBrowserMessages.ts
/**
 * See the NOTICE file distributed with this work for additional information
 * regarding copyright ownership.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

import type { GenomeBrowserLocationPayload } from './genomeBrowserLocation';
import type { TrackSummary } from './trackSummary';
import type { HotspotPayload, TranscriptsLozengePayload } from './hotspot';
import type { ZmenuPayload } from './zmenu';

export type GenomeBrowserMessageMap = {
  current_position: GenomeBrowserLocationPayload;
  target_position: GenomeBrowserLocationPayload;
  track_summary: { summary: TrackSummary[] };
  hotspot: HotspotPayload | TranscriptsLozengePayload | ZmenuPayload;
  error: unknown;
  'out-of-date': unknown; // NOTE: this is supposed to be temporary; hopefully, one day it will be merged with the rest of the error messages
};

export type CurrentPositionMessage = {
  type: 'current_position';
  payload: GenomeBrowserMessageMap['current_position'];
};

export type TargetPositionMessage = {
  type: 'target_position';
  payload: GenomeBrowserMessageMap['target_position'];
};

export type TrackSummaryMessage = {
  type: 'track_summary';
  payload: GenomeBrowserMessageMap['track_summary'];
};

export type HotspotMessage = {
  type: 'hotspot';
  payload: GenomeBrowserMessageMap['hotspot'];
};